Added a few lines to LinuxRCInput_Navio::termination_handler(int signum) in libraries/AP_HAL_Linux/RCInput_Navio.cpp to:
1. ignore signal 17
2. report which signal killed us

Signal 17 (SIGCHLD) is sent to the parent when the status of any child process changes. For example if you launch a process in UserCode.cpp, that ends normally, ArduCopter receives SIGCHLD and calls stop_dma(). This causes a segfault as SIGCHLD does not kill ArduCopter. Maybe the switch will be useful for speific action for other signals?

It is useful to know what signal killed us - for example to diagnose segfaults caused by SIGCHLD.